+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add help page
+- Add `activated` directive in `<Item>`
 
 ### Fixed
 - Make top-app-bar fixed